LPS-73793 Move *DestinationNames to constants package
