+++ modules/apps/adaptive-media/adaptive-media-api/src/main/java/com/liferay/adaptive/media/internal/constants/AdaptiveMediaDestinationNames.java
@@ -12,7 +12,7 @@
  * details.
  */
 
-package com.liferay.adaptive.media.internal.messaging;
+package com.liferay.adaptive.media.internal.constants;
 
 /**
  * @author Adolfo Pérez
