+++ modules/apps/adaptive-media/adaptive-media-api/src/main/java/com/liferay/adaptive/media/internal/processor/AdaptiveMediaAsyncProcessorImpl.java
@@ -14,7 +14,7 @@
 
 package com.liferay.adaptive.media.internal.processor;
 
-import com.liferay.adaptive.media.internal.messaging.AdaptiveMediaDestinationNames;
+import com.liferay.adaptive.media.internal.constants.AdaptiveMediaDestinationNames;
 import com.liferay.adaptive.media.internal.messaging.AdaptiveMediaProcessorCommand;
 import com.liferay.adaptive.media.processor.AdaptiveMediaAsyncProcessor;
 import com.liferay.portal.kernel.exception.PortalException;
