+++ modules/apps/adaptive-media/adaptive-media-image-api/src/main/java/com/liferay/adaptive/media/image/constants/AdaptiveMediaImageDestinationNames.java
@@ -12,7 +12,7 @@
  * details.
  */
 
-package com.liferay.adaptive.media.image.messaging;
+package com.liferay.adaptive.media.image.constants;
 
 /**
  * @author Alejandro Tardín
